workshops/security-for-developers/requirements.txt
attrs==22.1.0
aws-cdk-lib==2.38.1
cattrs==22.1.0
constructs==10.1.162
exceptiongroup==1.0.0rc8
jsii==1.71.0
publication==0.0.3
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
typeguard==2.13.3
typing-extensions==4.3.0